pyproject.toml
[project]
name = "open-pr-agent"
version = "0.1.0"
description = "FOSS AI PR review agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "openhands-sdk @ git+https://github.com/OpenHands/agent-sdk.git@main#subdirectory=openhands-sdk",
    "openhands-tools @ git+https://github.com/OpenHands/agent-sdk.git@main#subdirectory=openhands-tools",
]
